requirements.txt
streamlit
openai
requests
python-dotenv
streamlit-chat
google-generativeai
grpcio
anthropic